requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
psycopg2-binary==2.9.11
azure-storage-blob==12.28.0
azure-storage-queue==12.15.0
python-dotenv==1.1.0
python-multipart==0.0.22
pytest==9.0.2
httpx==0.28.1